gt58ltebmc: add overlay to restart preview after photo is taken by stock camera
Thanks to xda user dylux, he suggested that problem with stock camera not
showing the preview after a picture was taken was due to an overlay as per

"Hi, I did a bit of digging on the freeze issue. Looking at

https://github.com/LineageOS/android_packages_apps_Snap/blob/lineage-18.1/res/values/config.xml,

it seems to me that preview is not restarted after a picture is taken by default.
Therefore, adding the option "back_camera_restart_preview_onPictureTaken" and
"front_camera_restart_preview_onPictureTaken" to
overlay/packages/apps/Snap/res/values/config.xml should do the job.

I tested this on Snap with cameraV1, and initial result seems to suggest that this works."

https://forum.xda-developers.com/t/unofficial-rom-beta-lineageos-18-1-go-for-galaxy-tab-a-sm-t350-may-7-2022.4305903/page-5#post-86859253